requirements.txt
alabaster==0.7.12
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.10.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
idna==3.2
imagesize==1.2.0
Jinja2==3.0.2
kiwisolver==1.3.2
markdown-it-py==1.1.0
MarkupSafe==2.0.1
matplotlib==3.4.3
mdit-py-plugins==0.2.8
myst-parser==0.15.2
networkx==2.6.3
numpy==1.20.3
opencv-python==4.5.2.54
packaging==21.0
Pillow==8.2.0
pycodestyle==2.7.0
Pygments==2.10.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-jsonrpc-server==0.3.4
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rstcheck==3.3.1
six==1.16.0
snooty-lextudio==1.11.4.dev0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
watchdog==1.0.2
